Split `acctest` package into several files
